feat: Move nested attributes for interactions to a separate module
