+++ lib/plutonium/interaction/base.rb
@@ -26,6 +26,7 @@ module Plutonium
       include Plutonium::Definition::ConfigAttr
       include Plutonium::Definition::Presentable
       include Plutonium::Definition::NestedInputs
+      include Plutonium::Interaction::NestedAttributes
       # include Plutonium::Interaction::Concerns::WorkflowDSL
 
       class Form < Plutonium::UI::Form::Interaction; end
